[#3679] Fix dereference of null pointer in ssl_send_rods_msg
_msg_buf, _error_buf, and _stream_bbuf are parameters of the function
and are not null-checked before dereferencing when printing to log or
when attempting to write to socket.

This change adds the needed null checks to accompany the zero-length
buffer checks such that the result returned by the function is the
same for either case.

The buffer lengths "should" never be non-zero when _msg_buf,
_error_buf, and _stream_bbuf are null, so the null pointer
dereferences "should" never happen. This adds an extra layer of
protection.