[4331] New catalog library features
Adds a catalog utility to bind values to nanodbc statements. Also
adds a mapping between r_data_main columns and said catalog utilities
so that updating rows in the table becomes simpler.

Notably missing from the columns are the unused columns:
    data_resc_group
    data_resc_name
    data_resc_hier

Also adds utilities for redirecting to the catalog server and
determining if it is necessary to do so before performing catalog
operations. These should be helpful in the implementation of the
finalize API plugin being built.