[4331] Add finalize_data_object API plugin
Adds a new API plugin which finalizes a data object by atomically
updating all rows in r_data_main mapped to a particular data_id.

The new API endpoint takes a JSON string of the following form:

{
    "data_id": "XXXXX"
    "replicas": [
        {
            "before": {
                "data_id": "XXXXX",
                "coll_id": "XXXXX",
                ...
            }
            "after": {
                "data_id": "XXXXX",
                "coll_id": "XXXXX",
                ...
            }
        },
        ...
    ]
}

The API plugin takes the array of replicas and updates each corresponding
row in the catalog with the column values provided in the "after" piece.
The "before" piece is mainly in place for the caller to keep track of the
state of a given replica before any changes are made to it at the time of
opening.