[5104] Add data object proxy classes
Adds two new classes: data_object_proxy and replica_proxy

data_object_proxy serves to represent a data object in the code but
is merely a proxy to a dataObjInfo_t. replica_proxy serves to represent
a replica in the code but is merely a proxy to a dataObjInfo_t.

The major difference is the dataObjInfo_t* next member of the struct which
creates a linked list of replicas of a data object. The data_object_proxy
holds some information that applies to all replicas of a data object as
well as a std::vector of replica_proxy objects to represent the replicas.

Some new data members and convenience functions were created to hold
important data object information in the irods::file_object.