[_5312] Use finalize utilities and protect against L1desc
Removes/replaces uses of custom definitions of finalize logic in various
places and protects against some assumptions built in regarding the
L1desc and finalize.

Adds a selected hierarchy keyword to indicate the winner of the vote so
that fileModified can retain this information without the need to keep
the replica open. This prepares the ground for closing replicas before
finalizing.